update: 学習進捗率が更新されるように変更

+++ lib/presantation/home/view_model/home_screen_state.dart
@@ -5,6 +5,6 @@ part '../../../generated/presantation/home/view_model/home_screen_state.freezed.
 @freezed
 abstract class HomeScreenState with _$HomeScreenState {
   const factory HomeScreenState({
-    required int progress, // 進捗率
+    required double progress, // 進捗率
   }) = _HomeScreenState;
 }
